+++ packages/ember-data/lib/system/relationships.js
@@ -6,4 +6,4 @@
 require("ember-data/system/relationships/belongs_to");
 require("ember-data/system/relationships/has_many");
 require("ember-data/system/relationships/ext");
-require("ember-data/system/relationships/one_to_many_change");
+require("ember-data/system/relationships/relationship_change");
